+++ src/main/java/com/market/allForOneReview/domain/article/Controller/ReviewController.java
@@ -2,14 +2,12 @@ package com.market.allForOneReview.domain.article.Controller;
 
 import com.market.allForOneReview.domain.answer.AnswerForm;
 import com.market.allForOneReview.domain.article.Service.ReviewService;
-import com.market.allForOneReview.domain.article.entity.Category;
 import com.market.allForOneReview.domain.article.entity.Review;
 import com.market.allForOneReview.domain.article.entity.ReviewForm;
-import com.market.allForOneReview.domain.user.UserService;
 import com.market.allForOneReview.domain.user.entity.SiteUser;
+import com.market.allForOneReview.domain.user.service.UserService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-
 import org.springframework.data.domain.Page;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
@@ -18,7 +16,6 @@ import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import java.security.Principal;
-import java.util.List;
 
 @Controller
 @RequiredArgsConstructor
@@ -76,7 +73,7 @@ public class  ReviewController {
         if (bindingResult.hasErrors()) {
             return "create";
         }
-        SiteUser siteUser = this.userService.getUser(principal.getName());
+        SiteUser siteUser = this.userService.findByUsername(principal.getName());
         this.reviewService.create(reviewForm.getTitle(), reviewForm.getContentStory(), reviewForm.getContent(), reviewForm.getCategory(), reviewForm.getSubCategory(), siteUser);
 
         return "redirect:/review/sub";
